+++ bindings/angular2/examples/range.ts
@@ -17,9 +17,7 @@ import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
     <div class="background"></div>
     <div class="content">
       <div style="text-align: center; margin: 10px">
-        <!-- (input) is needed for immediate change detection on dragging range component.
-        -->
-        <ons-range id="range" [(value)]="value" (input)="value = $event.target.value"></ons-range><br>
+        <ons-range id="range" [(value)]="value"></ons-range><br>
         <ons-range modifier="material" [(value)]="value"></ons-range><br>
         <span id="value">
           {{ value }}